Change int to long

+++ src/SalesManager.java
@@ -1,14 +1,14 @@
 
 public class SalesManager {
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManager(int[] sales) {
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -17,9 +17,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = 1000;
-        for (int sale : sales) {
+    public long min() {
+        long min = 1000;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
